emery vpn orchestrator/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic-settings==2.10.1
python-dotenv==1.1.1
aiogram==3.26.0
python-dateutil==2.9.0.post0
httpx==0.28.1
paramiko==3.5.1
pytest==8.4.2
